Change class name to TorResources
