+++ library/resource-tor/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/resource/tor/internal/-JvmPlatform.kt
@@ -16,7 +16,7 @@
 package io.matthewnelson.kmp.tor.resource.tor.internal
 
 import io.matthewnelson.kmp.file.File
-import io.matthewnelson.kmp.tor.resource.tor.TorResource
+import io.matthewnelson.kmp.tor.resource.tor.TorResources
 import io.matthewnelson.kmp.tor.core.api.annotation.InternalKmpTorApi
 import io.matthewnelson.kmp.tor.core.resource.ImmutableMap
 import io.matthewnelson.kmp.tor.core.resource.OSHost
@@ -28,7 +28,7 @@ import io.matthewnelson.kmp.tor.core.resource.Resource
 @OptIn(InternalKmpTorApi::class)
 internal actual val RESOURCE_CONFIG: Resource.Config by lazy {
     Resource.Config.create {
-        val clazz = TorResource::class.java
+        val clazz = TorResources::class.java
 
         resource(ALIAS_GEOIP) {
             isExecutable = false
